requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
django-cors-headers==3.7.0
django-elasticsearch-dsl==7.1.4
djangorestframework==3.12.2
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
